requirements_trombinoscope.txt
Pillow>=10.0.0
python-docx>=1.0.0
docx2pdf>=0.1.8
pywin32>=305 ; platform_system=="Windows"
